FAI-614: configurable explanation section
