+++ ui-packages/packages/trusty/src/components/Organisms/OutcomeSwitch/OutcomeSwitch.scss
@@ -1,3 +1,3 @@
-.explanation-switch {
+.outcome-switch {
   max-width: 400px;
 }
