+++ ui-packages/packages/trusty/src/components/Templates/OutcomeDetails/tests/useOutcomeDetail.test.tsx
