+++ ui-packages/packages/trusty/src/components/Templates/OutcomeDetails/useOutcomeDetail.tsx
