+++ ui-packages/packages/trusty/src/components/Templates/OutcomeDetails/useSaliencies.tsx
